LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-replace-navigation-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'e.activation.entry')
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-replace-navigation.html?...

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-replace-navigation.html?new", "replace", "from", "pagehide"] length 6, got ["pageswap", "transition", "pagehide"] length 3

TIMEOUT pageswap on replace navigation from script Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-replace-navigation.html?new", "replace", "from", "pagehide"] length 6, got ["pageswap", "transition", "pagehide"] length 3

TIMEOUT pageswap on replace navigation from script Test timed out
